--- client/src/pages/Login.css
@@ -1,103 +1,0 @@
-* {
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
-}
-
-@font-face {
-  font-family: QanelasBold;
-  src: url(../fonts/Qanelas/QanelasBold.otf);
-}
-
-@font-face {
-  font-family: QanelasSemiBold;
-  src: url(../fonts/Qanelas/QanelasSemiBold.otf);
-}
-
-@font-face {
-  font-family: QanelasMedium;
-  src: url(../fonts/Qanelas/QanelasMedium.otf);
-}
-
-html,
-body {
-  height: 100%;
-  width: 100%;
-  font-family: "Arial", sans-serif;
-  font-size: 16px;
-  line-height: 1.6;
-  background: #fdeae5;
-  color: #333;
-}
-
-body {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  padding: 1rem;
-}
-
-a {
-  font-family: "QanelasSemiBold";
-}
-
-.login-container {
-  text-align: center;
-  width: 100%;
-}
-
-.header {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  margin: 0 0 100px 0;
-}
-
-.close-btn {
-  cursor: pointer;
-  width: 40px;
-}
-
-.login-heading {
-  display: flex;
-  font-family: "QanelasBold";
-  text-align: center;
-  align-items: center;
-  justify-content: center;
-  padding-left: 20px;
-}
-
-.signup-btn {
-  background-color: #ff4400;
-  border-radius: 12px;
-  border: none;
-  color: white;
-  width: 100px;
-  height: 40px;
-  text-align: center;
-  font-size: 16px;
-}
-
-.signup-btn:hover {
-  background-color: #d83c03;
-}
-
-.google-btn {
-  background-color: white;
-  border: 1px solid #b3b3b3;
-  border-radius: 12px;
-  font-family: "QanelasMedium";
-  font-size: 18px;
-  padding: 10px 40px;
-  cursor: pointer;
-}
-
-.google-btn:hover {
-  border: 1px solid #ff4400;
-}
-
-.google-btn img {
-  margin-right: 20px;
-  width: 25px;
-  height: auto;
-}
